Have sections fill up the viewport
Adding a height selector to the body allows the body
to take up the height of the entire viewport.

Giving each section a value of flex-grow: 1, will allow each
section to take up an equal amount of space in the viewport.

The sizes of these sections may change later, but we will use this
template as a starting point.